+++ src/mapping.js
@@ -6,5 +6,11 @@ export const mapping = {
     : 'https://us-central1-carllouismanuel-1e3a9.cloudfunctions.net/quotation',
   trackVisit: process.env.NODE_ENV === 'development'
     ? 'http://127.0.0.1:5001/carllouismanuel-1e3a9/us-central1/trackVisit'
-    : 'https://us-central1-carllouismanuel-1e3a9.cloudfunctions.net/trackVisit',
+    : 'https://trackvisit-fjb46y5zza-uc.a.run.app',
+  previewFeedback: process.env.NODE_ENV === 'development'
+    ? 'http://127.0.0.1:5001/carllouismanuel-1e3a9/us-central1/previewFeedback'
+    : 'https://us-central1-carllouismanuel-1e3a9.cloudfunctions.net/previewFeedback',
+  analyticsSummary: process.env.NODE_ENV === 'development'
+    ? 'http://127.0.0.1:5001/carllouismanuel-1e3a9/us-central1/analyticsSummary'
+    : 'https://us-central1-carllouismanuel-1e3a9.cloudfunctions.net/analyticsSummary',
 }